refactor(providers): reloca o X para src/providers/x/
processXCapture/buildPlatformDataX/computeSummaryX saem do controller para o módulo
coeso src/providers/x/ (ao lado do parser), expostos via xProvider no registry. É o
template que LinkedIn (#6) e Instagram (#7) seguem. Corrige import faltante de
ExportInstagramPost. Golden-master byte-idêntico, 27 testes verdes.

Refs #5
